refactor(-): :hammer: move viewset solicitacoes ativas e inativas e protocolo padrao
